requirements.txt
grpcio
grpcio-tools
trendspy
torch
TTS
transformers
pyngrok
google-generativeai